+++ packages/soda-hooks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-hooks
 
+## 6.8.1
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.26.0
+
 ## 6.8.0
 
 ### Minor Changes