+++ packages/soda-react-router/CHANGELOG.md
@@ -1,5 +1,11 @@
 # soda-react-router
 
+## 1.1.5
+
+### Patch Changes
+
+- soda-hooks@6.8.1
+
 ## 1.1.4
 
 ### Patch Changes